IncludeExcludeExtensionSpec: improve Javadoc, remove static block
Setting system property "spock.configuration" during class
initialisation is too late for Spock to pick up the config, so the
static block is pointless and misleads users to believe they can
actually configure the system this way.

Instead the Javadoc was amended by information how to set the system
property via '-D' JVM command-line argument when running a test from an
IDE (or from the command line, for that matter).